LayoutTests/contact-picker/contacts-select-while-presenting-picker-expected.txt

This test verifies that navigator.contacts.select fails when a contact picker is already being displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS contacts.length is 1
PASS contact.name[0] is "Person A"
PASS results[1] is "InvalidStateError"
PASS successfullyParsed is true

TEST COMPLETE
Show contacts
